+++ src/oci/data_science/models/create_project_details.py
@@ -1,0 +1,209 @@
+# coding: utf-8
+# Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
+
+
+from oci.util import formatted_flat_dict, NONE_SENTINEL, value_allowed_none_or_none_sentinel  # noqa: F401
+from oci.decorators import init_model_state_from_kwargs
+
+
+@init_model_state_from_kwargs
+class CreateProjectDetails(object):
+    """
+    Parameters needed to create a new project. Projects enable users to organize their data science work.
+    """
+
+    def __init__(self, **kwargs):
+        """
+        Initializes a new CreateProjectDetails object with values from keyword arguments.
+        The following keyword arguments are supported (corresponding to the getters/setters of this class):
+
+        :param display_name:
+            The value to assign to the display_name property of this CreateProjectDetails.
+        :type display_name: str
+
+        :param description:
+            The value to assign to the description property of this CreateProjectDetails.
+        :type description: str
+
+        :param compartment_id:
+            The value to assign to the compartment_id property of this CreateProjectDetails.
+        :type compartment_id: str
+
+        :param freeform_tags:
+            The value to assign to the freeform_tags property of this CreateProjectDetails.
+        :type freeform_tags: dict(str, str)
+
+        :param defined_tags:
+            The value to assign to the defined_tags property of this CreateProjectDetails.
+        :type defined_tags: dict(str, dict(str, object))
+
+        """
+        self.swagger_types = {
+            'display_name': 'str',
+            'description': 'str',
+            'compartment_id': 'str',
+            'freeform_tags': 'dict(str, str)',
+            'defined_tags': 'dict(str, dict(str, object))'
+        }
+
+        self.attribute_map = {
+            'display_name': 'displayName',
+            'description': 'description',
+            'compartment_id': 'compartmentId',
+            'freeform_tags': 'freeformTags',
+            'defined_tags': 'definedTags'
+        }
+
+        self._display_name = None
+        self._description = None
+        self._compartment_id = None
+        self._freeform_tags = None
+        self._defined_tags = None
+
+    @property
+    def display_name(self):
+        """
+        Gets the display_name of this CreateProjectDetails.
+        A user-friendly display name for the resource. Does not have to be unique, and can be modified. Avoid entering confidential information.
+
+
+        :return: The display_name of this CreateProjectDetails.
+        :rtype: str
+        """
+        return self._display_name
+
+    @display_name.setter
+    def display_name(self, display_name):
+        """
+        Sets the display_name of this CreateProjectDetails.
+        A user-friendly display name for the resource. Does not have to be unique, and can be modified. Avoid entering confidential information.
+
+
+        :param display_name: The display_name of this CreateProjectDetails.
+        :type: str
+        """
+        self._display_name = display_name
+
+    @property
+    def description(self):
+        """
+        Gets the description of this CreateProjectDetails.
+        A short blurb describing the project.
+
+
+        :return: The description of this CreateProjectDetails.
+        :rtype: str
+        """
+        return self._description
+
+    @description.setter
+    def description(self, description):
+        """
+        Sets the description of this CreateProjectDetails.
+        A short blurb describing the project.
+
+
+        :param description: The description of this CreateProjectDetails.
+        :type: str
+        """
+        self._description = description
+
+    @property
+    def compartment_id(self):
+        """
+        **[Required]** Gets the compartment_id of this CreateProjectDetails.
+        The `OCID`__ of the compartment to create the project in.
+
+        __ https://docs.cloud.oracle.com/iaas/Content/API/Concepts/identifiers.htm
+
+
+        :return: The compartment_id of this CreateProjectDetails.
+        :rtype: str
+        """
+        return self._compartment_id
+
+    @compartment_id.setter
+    def compartment_id(self, compartment_id):
+        """
+        Sets the compartment_id of this CreateProjectDetails.
+        The `OCID`__ of the compartment to create the project in.
+
+        __ https://docs.cloud.oracle.com/iaas/Content/API/Concepts/identifiers.htm
+
+
+        :param compartment_id: The compartment_id of this CreateProjectDetails.
+        :type: str
+        """
+        self._compartment_id = compartment_id
+
+    @property
+    def freeform_tags(self):
+        """
+        Gets the freeform_tags of this CreateProjectDetails.
+        Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace. See `Resource Tags`__.
+        Example: `{\"Department\": \"Finance\"}`
+
+        __ https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm
+
+
+        :return: The freeform_tags of this CreateProjectDetails.
+        :rtype: dict(str, str)
+        """
+        return self._freeform_tags
+
+    @freeform_tags.setter
+    def freeform_tags(self, freeform_tags):
+        """
+        Sets the freeform_tags of this CreateProjectDetails.
+        Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace. See `Resource Tags`__.
+        Example: `{\"Department\": \"Finance\"}`
+
+        __ https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm
+
+
+        :param freeform_tags: The freeform_tags of this CreateProjectDetails.
+        :type: dict(str, str)
+        """
+        self._freeform_tags = freeform_tags
+
+    @property
+    def defined_tags(self):
+        """
+        Gets the defined_tags of this CreateProjectDetails.
+        Defined tags for this resource. Each key is predefined and scoped to a namespace. See `Resource Tags`__.
+        Example: `{\"Operations\": {\"CostCenter\": \"42\"}}`
+
+        __ https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm
+
+
+        :return: The defined_tags of this CreateProjectDetails.
+        :rtype: dict(str, dict(str, object))
+        """
+        return self._defined_tags
+
+    @defined_tags.setter
+    def defined_tags(self, defined_tags):
+        """
+        Sets the defined_tags of this CreateProjectDetails.
+        Defined tags for this resource. Each key is predefined and scoped to a namespace. See `Resource Tags`__.
+        Example: `{\"Operations\": {\"CostCenter\": \"42\"}}`
+
+        __ https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm
+
+
+        :param defined_tags: The defined_tags of this CreateProjectDetails.
+        :type: dict(str, dict(str, object))
+        """
+        self._defined_tags = defined_tags
+
+    def __repr__(self):
+        return formatted_flat_dict(self)
+
+    def __eq__(self, other):
+        if other is None:
+            return False
+
+        return self.__dict__ == other.__dict__
+
+    def __ne__(self, other):
+        return not self == other